paytrack_python/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
python-dotenv
